Other changes to minimize compiler warnings.

+++ JVM/runtime/ArrayObject.h
@@ -103,7 +103,7 @@ void ArrayObject<Object*>::accept(ObjectVisitorInterface * visitor)
 {
 	visitor->visit(this);
 
-	for (int i = 0; i < this->size; i++)
+	for (size_t i = 0; i < this->size; i++)
 	{
 		visitor->visit((word)this->arrayData[i]);
 	}
